+++ pyo3-macros-backend/src/wrap.rs
@@ -22,25 +22,16 @@ impl Parse for WrapPyFunctionArgs {
     }
 }
 
-pub fn wrap_pyfunction_impl(args: WrapPyFunctionArgs) -> syn::Result<TokenStream> {
+pub fn wrap_pyfunction_impl(args: WrapPyFunctionArgs) -> TokenStream {
     let WrapPyFunctionArgs {
-        mut function,
+        function,
         comma_and_arg,
     } = args;
-    let span = function.span();
-    let last_segment = function
-        .segments
-        .last_mut()
-        .ok_or_else(|| err_spanned!(span => "expected non-empty path"))?;
-
-    last_segment.ident = function_wrapper_ident(&last_segment.ident);
-
-    let output = if let Some((_, arg)) = comma_and_arg {
-        quote! { #function(#arg) }
+    if let Some((_, arg)) = comma_and_arg {
+        quote! { #function::wrap(#function::DEF, #arg) }
     } else {
-        quote! { &|arg| #function(arg) }
-    };
-    Ok(output)
+        quote! { &|arg| #function::wrap(#function::DEF, arg) }
+    }
 }
 
 pub fn wrap_pymodule_impl(mut module_path: syn::Path) -> syn::Result<TokenStream> {
@@ -56,10 +47,6 @@ pub fn wrap_pymodule_impl(mut module_path: syn::Path) -> syn::Result<TokenStream
 
         &|py| unsafe { #module_path.make_module(py).expect("failed to wrap pymodule") }
     })
-}
-
-pub(crate) fn function_wrapper_ident(name: &Ident) -> Ident {
-    format_ident!("__pyo3_get_function_{}", name)
 }
 
 pub(crate) fn module_def_ident(name: &Ident) -> Ident {