+++ src/impl_.rs
@@ -13,5 +13,7 @@ pub(crate) mod not_send;
 #[doc(hidden)]
 pub mod pyclass;
 #[doc(hidden)]
+pub mod pyfunction;
+#[doc(hidden)]
 pub mod pymethods;
 pub mod pymodule;